feat: consolidate middleware
