+++ server/middleware/parsers.go
@@ -1,7 +1,10 @@
 package middleware
 
 import (
+	"encoding/json"
+	"errors"
 	"fmt"
+	"io"
 	"net/http"
 	"strings"
 
@@ -9,6 +12,31 @@ import (
 	"github.com/gonobo/jsonapi/v1/query"
 	"github.com/gonobo/jsonapi/v1/server"
 )
+
+func UseRequestBodyParser() server.Options {
+	return server.WithMiddleware(func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			defer r.Body.Close()
+
+			document := jsonapi.Document{}
+
+			if err := json.NewDecoder(r.Body).Decode(&document); errors.Is(err, io.EOF) {
+				// no document inside the payload; execute the next handler
+				next.ServeHTTP(w, r)
+				return
+			} else if err != nil {
+				// document parsing failed; return error to client
+				server.Error(w, fmt.Errorf("request document error: %w", err), http.StatusBadRequest)
+				return
+			}
+
+			// save document to context and continue
+			ctx := jsonapi.FromContext(r.Context())
+			ctx.Document = &document
+			next.ServeHTTP(w, jsonapi.RequestWithContext(r, ctx))
+		})
+	})
+}
 
 // FieldsetQueryParser is a function that parses the fieldset query parameters.
 type FieldsetQueryParser interface {
@@ -40,9 +68,9 @@ type PageQueryParser interface {
 	ParsePageQuery(*http.Request) (query.Page, error)
 }
 
-// UsePaginationQueryParser is a middleware that parses the sort parameters from the URL query and
+// UsePageQueryParser is a middleware that parses the sort parameters from the URL query and
 // stores them within the JSON:API context.
-func UsePaginationQueryParser(parser PageQueryParser) server.Options {
+func UsePageQueryParser(parser PageQueryParser) server.Options {
 	return server.WithMiddleware(func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			page, err := parser.ParsePageQuery(r)
